extract api response methods to a trait
